backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
snowflake-connector-python==3.12.3
pydantic==2.10.4
python-multipart==0.0.18
